+++ src/shared/options/doesRepositoryExist.test.ts
