+++ src/shared/options/getPrefillOrPromptedOption.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 
-import { getPrefillOrPromptedValue } from "./getPrefillOrPromptedValue.js";
+import { getPrefillOrPromptedOption } from "./getPrefillOrPromptedOption.js";
 
 const mockText = vi.fn();
 
@@ -15,7 +15,7 @@ describe("getPrefillOrPromptedValue", () => {
 	it("logs a pre-fill message when a first value already exists", async () => {
 		const existing = "existing value";
 
-		const actual = await getPrefillOrPromptedValue(existing, "");
+		const actual = await getPrefillOrPromptedOption(existing, "");
 
 		expect(actual).toEqual(existing);
 	});
@@ -25,7 +25,7 @@ describe("getPrefillOrPromptedValue", () => {
 
 		mockText.mockResolvedValue(expected);
 
-		const actual = await getPrefillOrPromptedValue(undefined, "");
+		const actual = await getPrefillOrPromptedOption(undefined, "");
 
 		expect(actual).toEqual(expected);
 	});
