+++ src/steps/writing/creation/dotGitHub/createDevelopment.ts
@@ -1,4 +1,4 @@
-import { InputValues } from "../../../../shared/readInputs.js";
+import { InputValues } from "../../../../shared/options/readOptions.js";
 
 export function createDevelopment(values: InputValues) {
 	const lintLines = [
