+++ src/steps/writing/creation/dotGitHub/index.ts
@@ -1,4 +1,4 @@
-import { InputValues } from "../../../../shared/readInputs.js";
+import { InputValues } from "../../../../shared/options/readOptions.js";
 import { createDotGitHubActions } from "./actions.js";
 import { createDotGitHubFiles } from "./createDotGitHubFiles.js";
 import { createDotGitHubIssueTemplate } from "./issueTemplate.js";
